EC2 test server.go split in multiple files by purpose.
